best day ever :)

+++ components/gs_effects/include/gs_effect_common_functions.h
@@ -8,6 +8,6 @@ int16_t max_arr_odd(int16_t* x, size_t size);
 
 int16_t max_arr(int16_t* x, size_t size);
 
-float sign(int x);
+float sign(float x);
 
 int16_t abs_val(int16_t x);
